--- package/kedro_viz/api/graphql/__init__.py
@@ -1,2 +1,0 @@
-# TODO: docstrings
-from .router import router
