--- package/kedro_viz/api/rest/__init__.py
@@ -1,1 +1,0 @@
-from .router import router
