+++ package/kedro_viz/api/rest/router.py
@@ -1,4 +1,4 @@
-"""`kedro_viz.api.router` defines routes and handling logic for the API."""
+"""`kedro_viz.api.rest.router` defines routes and handling logic for the API."""
 # pylint: disable=missing-function-docstring
 from fastapi import APIRouter
 from fastapi.responses import JSONResponse
