+++ package/tests/test_api/example_pipelines.json
@@ -1,0 +1,184 @@
+{
+  "nodes": [
+    {
+      "id": "f2e4bf0e",
+      "name": "Process Data",
+      "full_name": "process_data",
+      "tags": ["split"],
+      "pipelines": ["data_processing", "__default__"],
+      "type": "task",
+      "modular_pipelines": ["uk", "uk.data_processing"],
+      "parameters": { "uk.data_processing.train_test_split": 0.1 }
+    },
+    {
+      "id": "f0ebef01",
+      "name": "Params: Train Test Split",
+      "full_name": "params:uk.data_processing.train_test_split",
+      "tags": ["split"],
+      "pipelines": ["data_processing", "__default__"],
+      "type": "parameters",
+      "modular_pipelines": ["uk", "uk.data_processing"],
+      "layer": null,
+      "dataset_type": null
+    },
+    {
+      "id": "13399a82",
+      "name": "Raw Data",
+      "full_name": "uk.data_processing.raw_data",
+      "tags": ["split"],
+      "pipelines": ["data_processing", "__default__"],
+      "type": "data",
+      "modular_pipelines": ["uk", "uk.data_processing"],
+      "layer": "raw",
+      "dataset_type": "kedro.extras.datasets.pandas.csv_dataset.CSVDataSet"
+    },
+    {
+      "id": "0ecea0de",
+      "name": "Model Inputs",
+      "full_name": "model_inputs",
+      "tags": ["split", "train"],
+      "pipelines": ["data_processing", "__default__", "data_science"],
+      "type": "data",
+      "modular_pipelines": [],
+      "layer": "model_inputs",
+      "dataset_type": "kedro.extras.datasets.pandas.csv_dataset.CSVDataSet"
+    },
+    {
+      "id": "7b140b3f",
+      "name": "Train Model",
+      "full_name": "train_model",
+      "tags": ["train"],
+      "pipelines": ["__default__", "data_science"],
+      "type": "task",
+      "modular_pipelines": ["uk", "uk.data_science"],
+      "parameters": { "train_test_split": 0.1, "num_epochs": 1000 }
+    },
+    {
+      "id": "f1f1425b",
+      "name": "Parameters",
+      "full_name": "parameters",
+      "tags": ["train"],
+      "pipelines": ["__default__", "data_science"],
+      "type": "parameters",
+      "modular_pipelines": [],
+      "layer": null,
+      "dataset_type": null
+    },
+    {
+      "id": "d5a8b994",
+      "name": "Model",
+      "full_name": "uk.data_science.model",
+      "tags": ["train"],
+      "pipelines": ["__default__", "data_science"],
+      "type": "data",
+      "modular_pipelines": ["uk", "uk.data_science"],
+      "layer": null,
+      "dataset_type": "kedro.io.memory_dataset.MemoryDataSet"
+    },
+    {
+      "id": "uk.data_processing",
+      "name": "Data Processing",
+      "full_name": "uk.data_processing",
+      "tags": [],
+      "pipelines": ["__default__"],
+      "type": "modularPipeline",
+      "modular_pipelines": null,
+      "layer": null,
+      "dataset_type": null
+    },
+    {
+      "id": "uk",
+      "name": "Uk",
+      "full_name": "uk",
+      "tags": [],
+      "pipelines": ["__default__"],
+      "type": "modularPipeline",
+      "modular_pipelines": null,
+      "layer": null,
+      "dataset_type": null
+    },
+    {
+      "id": "uk.data_science",
+      "name": "Data Science",
+      "full_name": "uk.data_science",
+      "tags": [],
+      "pipelines": ["__default__"],
+      "type": "modularPipeline",
+      "modular_pipelines": null,
+      "layer": null,
+      "dataset_type": null
+    }
+  ],
+  "edges": [
+    { "source": "f1f1425b", "target": "7b140b3f" },
+    { "source": "0ecea0de", "target": "uk.data_science" },
+    { "source": "f0ebef01", "target": "uk" },
+    { "source": "7b140b3f", "target": "d5a8b994" },
+    { "source": "f0ebef01", "target": "uk.data_processing" },
+    { "source": "13399a82", "target": "f2e4bf0e" },
+    { "source": "uk.data_science", "target": "d5a8b994" },
+    { "source": "f1f1425b", "target": "uk" },
+    { "source": "0ecea0de", "target": "7b140b3f" },
+    { "source": "uk", "target": "d5a8b994" },
+    { "source": "f1f1425b", "target": "uk.data_science" },
+    { "source": "f2e4bf0e", "target": "0ecea0de" },
+    { "source": "uk.data_processing", "target": "0ecea0de" },
+    { "source": "f0ebef01", "target": "f2e4bf0e" },
+    { "source": "13399a82", "target": "uk" },
+    { "source": "13399a82", "target": "uk.data_processing" }
+  ],
+  "layers": ["raw", "model_inputs"],
+  "tags": [
+    { "id": "split", "name": "Split" },
+    { "id": "train", "name": "Train" }
+  ],
+  "pipelines": [
+    { "id": "__default__", "name": "Default" },
+    { "id": "data_science", "name": "Data Science" },
+    { "id": "data_processing", "name": "Data Processing" }
+  ],
+  "modular_pipelines": {
+    "__root__": {
+      "id": "__root__",
+      "name": "Root",
+      "inputs": [],
+      "outputs": [],
+      "children": [
+        { "id": "f1f1425b", "type": "parameters" },
+        { "id": "uk", "type": "modularPipeline" },
+        { "id": "0ecea0de", "type": "data" }
+      ]
+    },
+    "uk.data_processing": {
+      "id": "uk.data_processing",
+      "name": "Data Processing",
+      "inputs": ["13399a82", "f0ebef01"],
+      "outputs": ["0ecea0de"],
+      "children": [
+        { "id": "13399a82", "type": "data" },
+        { "id": "f2e4bf0e", "type": "task" }
+      ]
+    },
+    "uk": {
+      "id": "uk",
+      "name": "Uk",
+      "inputs": ["13399a82", "f0ebef01", "f1f1425b"],
+      "outputs": ["d5a8b994"],
+      "children": [
+        { "id": "uk.data_science", "type": "modularPipeline" },
+        { "id": "uk.data_processing", "type": "modularPipeline" }
+      ]
+    },
+    "uk.data_science": {
+      "id": "uk.data_science",
+      "name": "Data Science",
+      "inputs": ["0ecea0de", "f1f1425b"],
+      "outputs": ["d5a8b994"],
+      "children": [
+        { "id": "d5a8b994", "type": "data" },
+        { "id": "7b140b3f", "type": "task" }
+      ]
+    }
+  },
+  "selected_pipeline": "__default__"
+}
